seperate account component to another app
