--- packages/forms.py
@@ -1,20 +1,0 @@
-from django.contrib.auth.models import User
-from django import forms
-from django.core.exceptions import ValidationError
-from django.contrib.auth import authenticate
-from crispy_forms.helper import FormHelper
-from django.shortcuts import redirect
-from django.utils.html import strip_tags
-from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
-
-class LoginForm(AuthenticationForm):
-    class Meta:
-        model = User
-        fields = ("username", "password",)
-
-class RegisterForm(UserCreationForm):
-    email = forms.EmailField(max_length=254, help_text='Required. Inform a valid email address.')
-
-    class Meta:
-        model = User
-        fields = ('username', 'email', 'password1', 'password2', )
